Linux: codelite is now using wx29 under Linux as well The codelite deb will now include wxWidgets shared libraries (until the official repos of the various distros will distribute them) As part of the change, codelite repository http://www.codelite.co.uk/ubuntu/ and http://www.codelite.co.uk/debian/ will offer wxWidgets packages for building codelite) Fixed sluggish startup of codelite under Linux with 2.94 Fixed assertion in TagsManager
git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@5823 9da81c78-c036-0410-9e1f-a2b0375e4b5a

+++ CodeLite/ctags_manager.cpp
@@ -366,14 +366,14 @@ void TagsManager::SourceToTags(const wxFileName& source, wxString& tags)
 
     // connect to the indexer
     if (!client.connect()) {
-        wxPrintf(wxT("Failed to connect to indexer ID %d!\n"), wxGetProcessId());
+        wxPrintf(wxT("Failed to connect to indexer ID %d!\n"), (int)wxGetProcessId());
         return;
     }
 
 
     // send the request
     if ( !clIndexerProtocol::SendRequest(&client, req) ) {
-        wxPrintf(wxT("Failed to send request to indexer ID [%d]\n"), wxGetProcessId());
+        wxPrintf(wxT("Failed to send request to indexer ID [%d]\n"), (int)wxGetProcessId());
         return;
     }
